rename + delete files
