+++ src/hyperactive/optimizers/_constraint.py
