+++ src/hyperactive/optimizers/_dictionary.py
