+++ src/hyperactive/optimizers/_optimizer_attributes.py
